# [Fix #1032] Better/faster performance when querying certificates on OS X