--- a/osquery/tables/system/darwin/keychain.h
+++ b/osquery/tables/system/darwin/keychain.h
@@ -15,62 +15,56 @@
 #include <CoreFoundation/CoreFoundation.h>
 #include <Security/Security.h>
 
+#include <openssl/x509v3.h>
+#include <openssl/bn.h>
+#include <openssl/asn1.h>
+#include <openssl/x509.h>
+#include <openssl/x509_vfy.h>
+#include <openssl/pem.h>
+#include <openssl/bio.h>
+
 #include "osquery/core/conversions.h"
 
 namespace osquery {
 namespace tables {
 
-typedef std::string (*PropGenerator)(const CFDataRef&);
-
-/// A helper data structure to apply a decode generator to property.
-struct CertProperty {
-  /// Property key.
-  CFTypeRef type;
-  /// Generator function.
-  PropGenerator generate;
-};
-
 extern const std::vector<std::string> kSystemKeychainPaths;
 extern const std::vector<std::string> kUserKeychainPaths;
+
+// The flags are defined in openssl/x509v3.h,
+// and its keys in crypto/x509v3/v3_bitst.c
+// clang-format off
+const std::map<unsigned long, std::string> kKeyUsageFlags = {
+    {0x0001, "Encipher Only"},
+    {0x0002, "CRL Sign"},
+    {0x0004, "Key Cert Sign"},
+    {0x0008, "Key Agreement"},
+    {0x0010, "Data Encipherment"},
+    {0x0020, "Key Encipherment"},
+    {0x0040, "Non Repudiation"},
+    {0x0080, "Digital Signature"},
+    {0x8000, "Decipher Only"}
+};
+// clang-format on
 
 void genKeychains(const std::string& path, CFMutableArrayRef& keychains);
 std::string getKeychainPath(const SecKeychainItemRef& item);
 
 /// Certificate property parsing functions.
-std::string genKIDProperty(const CFDataRef& kid);
-std::string genCommonNameProperty(const CFDataRef& constraints);
-std::string genAlgProperty(const CFDataRef& alg);
-std::string genCAProperty(const CFDataRef& ca);
+std::string genKIDProperty(const unsigned char* data, int len);
+std::string genAlgProperty(const X509* cert);
+std::string genCommonName(X509* cert);
+time_t genEpoch(ASN1_TIME* time);
 
-/// Not a property generator, do not use in kCertificateProperties.
-std::string genSHA1ForCertificate(const SecCertificateRef& ca);
-
-CFDataRef CreatePropertyFromCertificate(const SecCertificateRef& cert,
-                                        const CFTypeRef& oid);
-bool CertificateIsCA(const SecCertificateRef& cert);
+std::string genSHA1ForCertificate(const CFDataRef& raw_cert);
+bool CertificateIsCA(X509* cert);
 
 /// Generate a list of keychain items for a given item type.
 CFArrayRef CreateKeychainItems(const std::set<std::string>& paths,
                                const CFTypeRef& item_type);
 
 std::set<std::string> getKeychainPaths();
-
-// From SecCertificatePriv.h
-typedef uint32_t SecKeyUsage;
-enum {
-  kSecKeyUsageUnspecified = 0,
-  kSecKeyUsageDigitalSignature = 1 << 0,
-  kSecKeyUsageNonRepudiation = 1 << 1,
-  kSecKeyUsageContentCommitment = 1 << 1,
-  kSecKeyUsageKeyEncipherment = 1 << 2,
-  kSecKeyUsageDataEncipherment = 1 << 3,
-  kSecKeyUsageKeyAgreement = 1 << 4,
-  kSecKeyUsageKeyCertSign = 1 << 5,
-  kSecKeyUsageCRLSign = 1 << 6,
-  kSecKeyUsageEncipherOnly = 1 << 7,
-  kSecKeyUsageDecipherOnly = 1 << 8,
-  kSecKeyUsageCritical = 1 << 31,
-  kSecKeyUsageAll = 0x7FFFFFFF
-};
+std::string genKeyUsage(unsigned long flag);
+std::string genHumanReadableDateTime(ASN1_TIME* time);
 }
 }