--- a/osquery/tables/system/darwin/tests/certificates_tests.cpp
+++ b/osquery/tables/system/darwin/tests/certificates_tests.cpp
@@ -26,59 +26,51 @@ class CACertsTests : public ::testing::Test {
 
     raw = base64Decode(getCACertificateContent());
     data =
-        CFDataCreate(nullptr, (const UInt8*)raw.c_str(), (CFIndex)raw.size());
+        CFDataCreate(nullptr, (const UInt8 *)raw.c_str(), (CFIndex)raw.size());
     cert = SecCertificateCreateWithData(nullptr, data);
+    cert_der_data = SecCertificateCopyData(cert);
+    auto bytes = CFDataGetBytePtr(cert_der_data);
+    x_cert = d2i_X509(nullptr, &bytes, CFDataGetLength(cert_der_data));
+
     CFRelease(data);
   }
 
   virtual void TearDown() {
-    if (cert != NULL) {
+    if (cert != nullptr) {
       CFRelease(cert);
+    }
+    if (cert_der_data != nullptr) {
+      CFRelease(cert_der_data);
+    }
+    if (x_cert != nullptr) {
+      X509_free(x_cert);
     }
   }
 
   SecCertificateRef cert;
+  CFDataRef cert_der_data;
+  X509 *x_cert;
 };
 
 TEST_F(CACertsTests, test_certificate_sha1) {
   std::string sha1;
-  sha1 = genSHA1ForCertificate(cert);
+  sha1 = genSHA1ForCertificate(cert_der_data);
 
   EXPECT_EQ("f149bae28e3c754ff4bb062b2c1b8bac81b8783e", sha1);
 }
 
 TEST_F(CACertsTests, test_certificate_properties) {
-  CFDataRef property;
-  CFTypeRef oid;
-  std::string prop_string;
+  EXPECT_EQ("localhost.localdomain", genCommonName(x_cert));
 
-  oid = kSecOIDCommonName;
-  property = CreatePropertyFromCertificate(cert, oid);
-  prop_string = genCommonNameProperty(property);
+  X509_check_ca(x_cert);
+  auto skid = genKIDProperty(x_cert->skid->data, x_cert->skid->length);
+  EXPECT_EQ("f2b99b00e0ee60d57c426ce3e64e3fdc6f6411c0", skid);
 
-  EXPECT_EQ("localhost.localdomain", prop_string);
-  CFRelease(property);
+  auto not_before = std::to_string(genEpoch(X509_get_notBefore(x_cert)));
+  EXPECT_EQ("1408475536", not_before);
 
-  oid = kSecOIDSubjectKeyIdentifier;
-  property = CreatePropertyFromCertificate(cert, oid);
-  prop_string = genKIDProperty(property);
-
-  EXPECT_EQ("f2b99b00e0ee60d57c426ce3e64e3fdc6f6411c0", prop_string);
-  CFRelease(property);
-
-  oid = kSecOIDX509V1ValidityNotBefore;
-  property = CreatePropertyFromCertificate(cert, oid);
-  prop_string = stringFromCFNumber(property);
-
-  EXPECT_EQ("430168336", prop_string);
-  CFRelease(property);
-
-  oid = kSecOIDBasicConstraints;
-  property = CreatePropertyFromCertificate(cert, oid);
-  prop_string = genCAProperty(property);
-
-  EXPECT_EQ("1", prop_string);
-  CFRelease(property);
+  auto ca = (CertificateIsCA(x_cert)) ? "1" : "0";
+  EXPECT_EQ("1", ca);
 }
 }
 }