@osqueryer 2015-07-11
Can one of the osquery admins reply with "ok to test" to kick off a build...