@theopolis 2015-07-11
Also, if you want to squash deprecated warnings for the OS-provided OpenSSL you can wrap the API calls with [`OSQUERY_USE_DEPRECATED`](https://github.com/facebook/osquery/blob/master/include/osquery/tables.h#L27). We can also start using the Homebrew-provided OpenSSL and statically link in the symbols we need from libcrypto. It's a cardinal sin on other platforms, but we're already damned on OS X.