@sharvilshah 2015-07-11
I would wrap OpenSSL API calls with `OSQUERY_USE_DEPRECATED` if there were only a handful of them, but there are about ~12-15 calls, which I don't think makes it worth it. But I don't feel strongly about it either way, your call.

I did try linking it newer Homebrew provided OpenSSL first, but couldn't find a straightforward way to link missing/undefined symbols:

```
Linking CXX executable osqueryd
Undefined symbols for architecture x86_64:
  "_ERR_remove_thread_state", referenced from:
      boost::asio::ssl::detail::openssl_init_base::do_init::~do_init() in libosquery_additional.a(tls.cpp.o)
      boost::asio::ssl::detail::openssl_init_base::do_init::~do_init() in libosquery_additional.a(tls.cpp.o)
      boost::asio::ssl::detail::openssl_init_base::do_init::~do_init() in libosquery_additional.a(tls.cpp.o)
      boost::asio::ssl::detail::openssl_init_base::do_init::~do_init() in libosquery_additional.a(tls.cpp.o)
      boost::asio::ssl::detail::openssl_init_base::do_init::~do_init() in libcppnetlib-client-connections.a(client.cpp.o)
  "_TLSv1_1_client_method", referenced from:
      boost::asio::ssl::context::context(boost::asio::ssl::context_base::method) in libcppnetlib-client-connections.a(client.cpp.o)
  "_TLSv1_1_method", referenced from:
      boost::asio::ssl::context::context(boost::asio::ssl::context_base::method) in libcppnetlib-client-connections.a(client.cpp.o)
  "_TLSv1_1_server_method", referenced from:
      boost::asio::ssl::context::context(boost::asio::ssl::context_base::method) in libcppnetlib-client-connections.a(client.cpp.o)
  "_TLSv1_2_client_method", referenced from:
      boost::asio::ssl::context::context(boost::asio::ssl::context_base::method) in libcppnetlib-client-connections.a(client.cpp.o)
  "_TLSv1_2_method", referenced from:
      boost::asio::ssl::context::context(boost::asio::ssl::context_base::method) in libcppnetlib-client-connections.a(client.cpp.o)
  "_TLSv1_2_server_method", referenced from:
      boost::asio::ssl::context::context(boost::asio::ssl::context_base::method) in libcppnetlib-client-connections.a(client.cpp.o)
ld: symbol(s) not found for architecture x86_64
clang: error: linker command failed with exit code 1 (use -v to see invocation)
make[3]: *** [osquery/osqueryd] Error 1
make[2]: *** [osquery/CMakeFiles/daemon.dir/all] Error 2
make[1]: *** [all] Error 2
make: *** [all] Error 2
```