@theopolis 2015-07-11
I'd guess those link errors stem from using the Homebrew-provided OpenSSL's include headers but the Apple-provided libcrypto.so library for linking.

If it was using the Homebrew-OpenSSL's headers you'll have >TLS1.0 support and thus enable those ASIO/SSL features as well as remove the deprecated warnings which come from header macros.

To completely switch to Homebrew's version we'd have to modify the [provision scripts slightly](https://github.com/facebook/osquery/blob/master/tools/provision/darwin.sh#L29) to add openssl, [assure](https://github.com/Homebrew/homebrew/blob/master/Library/Formula/openssl.rb#L32) a static version of libcrypto(.a) was installed, and change [CMake's preference](https://github.com/facebook/osquery/blob/master/CMakeLists.txt#L273) for static/dynamic. It's a bit of work, but definitely worth it.