@sharvilshah 2015-07-12
Yes, you are right, I used Homebrew-provided OpenSSL headers, but linked with Apple's libcrypto.so.
I had already spent way too much time figuring out OpenSSL APIs for relatively simple things and so just stuck with Apple provided OpenSSL.

Thanks for the pointers, those should be very helpful. I will try taking a crack at it later today/tomorrow and will probably still need to ping you on IRC.