@theopolis 2015-07-12
Perfect, we'll figure out a solution for OpenSSL going forward in a follow up set of PRs/issues/commits. I want some time to test 10.11 and the libressl included for openssh too!