@theopolis 2015-07-12
This is ready to merge as soon as we figure out how to handle the CA checking. Either we count any non-0 returns from `X509_check_ca` as true or emit the exact return integer. Your call!