@sharvilshah 2015-07-12
Okay, updated so that `X509_check_ca`'s return code  > 0  is true.